V2dn3org (#25)
* Add 3orgs to v2 dev-net except bootstrap.sh

* Make sure testing with tester container also clean-up after fail tests

* Refactor argument parsing for dev-net scripts

* Add 3orgs to v2 dev-net

* Add change log

* Work around lint issue in Github

* Fix lint rules violation not appeared previously

+++ deployments/dev-net/cleanup.sh
@@ -1,13 +1,56 @@
 #!/usr/bin/env bash
 
 # Cleaup the environment
+# Usage: cleanup.sh [-d | -R | --remove-cc-images] [docker-compose file]
 
 . ./scripts/setup.sh
 
-COMPOSE=$COMPOSE_4_3ORG
-if [ $# -eq 1 ]; then
-  COMPOSE=$1
+NEEDSUDO=1
+if [[ "$OSTYPE" == "darwin"* ]]; then
+  NEEDSUDO=0
 fi
+
+COMPOSE=$COMPOSE_4_3ORG
+CLEAN_CC_IMG=0
+case $# in
+  0)
+    ;;
+  1|2)
+    case $1 in
+      -d)
+        if [ $# -gt 1 ]; then
+          printMessage "Usage: cleanup.sh {-R | --remove-cc-images} {[docker-compose file]}; cleanup" 2
+        fi
+        ;;
+      -R|--remove-cc-images)
+        CLEAN_CC_IMG=1
+        if [ $# -eq 2 ]; then
+          case $2 in
+            -f*)
+              COMPOSE=$2
+              ;;
+            *)
+              printMessage "Usage: cleanup.sh {-R | --remove-cc-images} {[docker-compose file]}; cleanup" 5
+              ;;
+          esac
+        fi
+        ;;
+      -f*)
+        if [ $# -eq 1 ]; then
+          COMPOSE=$1
+        else
+          printMessage "Usage: cleanup.sh {-R | --remove-cc-images} {[docker-compose file]}; cleanup" 1
+        fi
+        ;;
+      *)
+        printMessage "Usage: cleanup.sh {-R | --remove-cc-images} {[docker-compose file]}; cleanup" 3
+        ;;
+    esac
+    ;;
+  *)
+    printMessage "Usage: cleanup.sh {-R | --remove-cc-images} {[docker-compose file]}; cleanup" 4
+    ;;
+esac
 
 LOGSPOUT=`docker ps -a | grep logspout`
 if [ ! -z "$LOGSPOUT" ]; then
@@ -25,14 +68,31 @@ fi
 docker volume prune -f
 docker network prune -f
 
-sudo rm -rf $VOLUME
-printMessage "Remove ${VOLUME}" $?
+if [ $NEEDSUDO -eq 1 ]; then
+  sudo rm -rf $VOLUME
+  printMessage "Remove ${VOLUME}" $?
 
-sudo rm -rf $ARTIFACTS
-printMessage "Remove ${ARTIFACTS}" $?
+  sudo rm -rf $ARTIFACTS
+  printMessage "Remove ${ARTIFACTS}" $?
+else
+  rm -fr $VOLUME
+  printMessage "Remove ${VOLUME}" $?
 
-docker rmi $(docker images -qf "dangling=true")
+  rm -rf $ARTIFACTS
+  printMessage "Remove ${ARTIFACTS}" $?
+fi
+
+DANGLE=`docker images -qf "dangling=true"`
+if [ ! -z "$DANGLE" ]; then
+  echo "Cleaning up dangling docker images..."
+  docker rmi $(docker images -qf "dangling=true")
+fi
 
 # Cleanup chaincode images
-#docker rmi -f $(docker images | grep eventstore | awk '{print $3}')
-#docker rmi -f $(docker images | grep privatedata | awk '{print $3}')
+if [ $CLEAN_CC_IMG -eq 1 ]; then
+  CHAIN=`docker images -qf "reference=*eventstore*"`
+  if [ ! -z "$CHAIN" ]; then
+    echo "Cleaning up chaincode docker images..."
+    docker rmi $(docker images -qf "reference=*eventstore*")
+  fi
+fi
